+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [5.3.2 (unreleased)](https://github.com/kdeldycke/workflows/compare/v5.3.1...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [5.3.1 (2026-01-24)](https://github.com/kdeldycke/workflows/compare/v5.3.0...v5.3.1)
 
 - Add `gha-utils sponsor-label` command to label issues/PRs from GitHub sponsors.